requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
Django==3.2
django-environ==0.8.1
fonttools==4.28.5
GDAL==3.0.4
geoserver-rest==2.1.4
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
Pygments==2.11.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
Werkzeug==2.0.2
